Fixing double for price values BigDecimal used instead

+++ src/main/java/com/andersenlab/hotel/usecase/CalculateClientStayCurrentPriceUseCase.java
@@ -1,8 +1,9 @@
 package com.andersenlab.hotel.usecase;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface CalculateClientStayCurrentPriceUseCase {
 
-    double calculatePrice(UUID id);
+    BigDecimal calculatePrice(UUID id);
 }
